Переносит функцию для валидации хэштегов и получения сообщений для ошибки в отдельный модуль, подключает его в image-upload-form.js
